+++ tests/testthat/test-vega-embed.R
@@ -1,5 +1,3 @@
-context("vega-embed")
-
 test_that("validation works", {
 
   # renderer