+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## Bug fixes
 
-* Unique identifier used in attaching data from local files; `elementId` is used if provided, if not an md5 hash of the files is used. (#125)
+* Unique identifier used in attaching data from local files; `elementId` is used if provided, if not an md5 hash of the files is used.
+  This would be useful for knitting documents, but attaching local data works only for interactive analysis, not for knitr environments.
+  Hence this "fix" is anticipation of other enchancements. (#125)
 
 * `vega_embed()`: