+++ NEWS.md
@@ -2,14 +2,22 @@
 
 * Upgrade to Vega-Lite 4.17.0, Vega 5.17.0, vega-embed 6.12.2. 
 
+* Unique identifier used in attaching data from local files; `elementId` is used if provided, if not an md5 hash of the files is used.
+  This would be useful for knitting documents, but attaching local data works only for interactive analysis, not for knitr environments.
+  Hence this "fix" is anticipation of other enchancements. (#125)
+
 * `vega_embed()`: 
 
   - the `actions` argument lets you specify `export$png` and `export$svg`. (#115)
   
   - the bulk of the options for [vega-embed](https://github.com/vega/vega-embed) can be supplied as named arguments using `...` rather than as explicit arguments to `vega_embed()`. 
   This will make it easier to support future enhancements to vega-embed.
+  
+* `vegawidget()`: keep precision in JSON representation for accurate rendering of small values (< 0.001). (#130, @datapixie)
 
-* repair `knit_print()` by exporting the S3 method. (#132)
+* `knit_print()`: repair by exporting the S3 method. (#132)
+
+* `as_vegaspec()`: adds `encoding` argument (default `"UTF-8"`) for `character` method, for files and URLs. (#117, with @g3o2)
 
 # vegawidget 0.3.1