+++ NEWS.md
@@ -1,4 +1,4 @@
-# vegawidget 0.3.1 (development)
+# vegawidget 0.3.2
 
 * Upgrade to Vega-Lite 4.17.0, Vega 5.17.0, vega-embed 6.12.2.